GUI/requirements.txt
admet-ai==1.2.0
alabaster==0.7.12
archspec==0.2.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
boltons==23.0.0
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
chemfunc==1.0.4
chemprop==1.6.1
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
defusedxml==0.7.1
descriptastorus==2.6.1
distro==1.8.0
docopt==0.6.2
docstring-parser==0.15
docutils==0.18.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.25.0
fsspec==2023.10.0
future==0.18.3
fuzzywuzzy==0.18.0
gmpy2==2.1.2
greenlet==3.0.1
gunicorn==21.2.0
huggingface-hub==0.20.2
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
importlib-metadata==7.0.1
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.7
mc-bin-client==1.0.1
mistune==3.0.2
# mkl-random==1.2.4
# mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.1
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.0.0
prompt-toolkit==3.0.43
protobuf==3.20.3
pure-eval==0.2.2
py4j==0.10.9.7
pycairo==1.23.0
pycosat==0.6.6
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
PySocks==1.7.1
PyTDC==0.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
QDarkStyle==3.2.3
qt-material==2.14
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtPy==2.4.1
rdkit==2023.9.4
rdkit-pypi==2022.9.5
referencing==0.35.1
reportlab==3.5.67
requests==2.31.0
rpds-py==0.18.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.13.1
setuptools==68.2.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.51
stack-data==0.6.3
sympy==1.12
tap==0.2
tensorboardX==2.2
threadpoolctl==2.2.0
tinycss2==1.3.0
torch==2.1.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
truststore==0.8.0
typed-argument-parser==1.9.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
win-inet-pton==1.1.0
xarray==2023.6.0
yarg==0.1.9
zipp==3.17.0
zstandard==0.19.0
